Added Swashbuckle and config for testing in Swagger. Added CRUD in AnimalsController for specific users. Added NameIdentifier for auth

+++ animal-finder-backend/AnimalFinderBackend/AnimalFinderBackend/Controllers/AnimalsController.cs
@@ -9,6 +9,7 @@ using AnimalFinderBackend.Data;
 using AnimalFinderBackend.Models;
 using System.Security.Claims;
 using AnimalFinderBackend.DTOs;
+using Microsoft.AspNetCore.Authorization;
 
 namespace AnimalFinderBackend.Controllers
     {
@@ -44,15 +45,28 @@ namespace AnimalFinderBackend.Controllers
             return animal;
             }
 
+
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<Animal>> PostAnimal(AnimalDto animalDto)
             {
+            //Getting Id from token
             var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
 
-            if (userId == null)
+
+
+            //Control if Id is empty
+            if (string.IsNullOrEmpty(userId))
                 {
                 return Unauthorized();
                 }
+            //Getting user from db
+            var user = await _context.Users.FindAsync(userId);
+            if (user == null)
+                {
+                return Unauthorized();
+                }
+
 
             var animal = new Animal
                 {
@@ -63,7 +77,8 @@ namespace AnimalFinderBackend.Controllers
                 DateOfDisappearance = animalDto.DateOfDisappearance,
                 ImageUrl = animalDto.ImageUrl,
                 UserId = userId,
-                DateAdded = DateTime.UtcNow
+                DateAdded = DateTime.UtcNow,
+
                 };
 
             _context.Animals.Add(animal);
@@ -73,16 +88,56 @@ namespace AnimalFinderBackend.Controllers
             }
 
 
-
-        // PUT: api/Animals/5
-        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
-        [HttpPut("{id}")]
-        public async Task<IActionResult> PutAnimal(int id, Animal animal)
+        //Getting animals for a specific user
+        [Authorize]
+        [HttpGet("user")]
+        public async Task<ActionResult<IEnumerable<Animal>>> GetUserAnimals()
             {
-            if (id != animal.AnimalId)
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+
+            if (string.IsNullOrEmpty(userId))
                 {
-                return BadRequest();
+                return Unauthorized();
                 }
+
+            var animals = await _context.Animals.Where(a => a.UserId == userId).ToListAsync();
+
+            if (animals.Count == 0)
+                {
+                return NotFound("You do not have any animals listed.");
+                }
+
+            return Ok(animals);
+            }
+
+        //Updating an animal for a specific user
+        [Authorize]
+        [HttpPut("{id}")]
+        public async Task<IActionResult> PutUserAnimal(int id, AnimalDto animalDto)
+            {
+            //getting id from token
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+
+            if (string.IsNullOrEmpty(userId))
+                {
+                return Unauthorized();
+                }
+
+            //Getting specific animal from id
+            var animal = await _context.Animals.FindAsync(id);
+
+            //If animal not found
+            if (animal == null || animal.UserId != userId)
+                {
+                return NotFound();
+                }
+
+            animal.Type = animalDto.Type;
+            animal.Name = animalDto.Name;
+            animal.Description = animalDto.Description;
+            animal.Neighborhood = animalDto.Neighborhood;
+            animal.DateOfDisappearance = animalDto.DateOfDisappearance;
+            animal.ImageUrl = animalDto.ImageUrl;
 
             _context.Entry(animal).State = EntityState.Modified;
 
@@ -94,7 +149,7 @@ namespace AnimalFinderBackend.Controllers
                 {
                 if (!AnimalExists(id))
                     {
-                    return NotFound();
+                    return NotFound("The animal does not exist");
                     }
                 else
                     {
@@ -105,25 +160,33 @@ namespace AnimalFinderBackend.Controllers
             return NoContent();
             }
 
-        //// POST: api/Animals
-        //// To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
-        //[HttpPost]
-        //public async Task<ActionResult<Animal>> PostAnimal(Animal animal)
-        //    {
-        //    _context.Animals.Add(animal);
-        //    await _context.SaveChangesAsync();
 
-        //    return CreatedAtAction("GetAnimal", new { id = animal.AnimalId }, animal);
-        //    }
 
-        // DELETE: api/Animals/5
+
+
+
+        //Deleting a specific animal for a specific user
+        [Authorize]
         [HttpDelete("{id}")]
-        public async Task<IActionResult> DeleteAnimal(int id)
+        public async Task<IActionResult> DeleteUserAnimal(int id)
             {
+
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+
+            if (string.IsNullOrEmpty(userId))
+                {
+                return Unauthorized();
+                }
+
             var animal = await _context.Animals.FindAsync(id);
             if (animal == null)
                 {
-                return NotFound();
+                return NotFound("The animal was not found.");
+                }
+
+            if (animal.UserId != userId)
+                {
+                return StatusCode(403, "You do not have the right to delete this animal");
                 }
 
             _context.Animals.Remove(animal);
@@ -131,6 +194,7 @@ namespace AnimalFinderBackend.Controllers
 
             return NoContent();
             }
+
 
         private bool AnimalExists(int id)
             {
